Fix tests to align with refactor
- Move some classes out of models to operations
