+++ blobxfer/models/crypto.py
@@ -32,14 +32,9 @@ from builtins import (  # noqa
 # stdlib imports
 import base64
 import collections
-import enum
 import hashlib
 import hmac
 import json
-try:
-    import queue
-except ImportError:  # noqa
-    import Queue as queue
 # non-stdlib imports
 # local imports
 import blobxfer.models.offload
@@ -47,7 +42,7 @@ import blobxfer.operations.crypto
 import blobxfer.util
 
 # encryption constants
-_AES256_BLOCKSIZE_BYTES = 16
+AES256_BLOCKSIZE_BYTES = 16
 
 # named tuples
 EncryptionBlobxferExtensions = collections.namedtuple(
@@ -316,59 +311,3 @@ class EncryptionMetadata(object):
             return hmac.new(self._signkey, digestmod=hashlib.sha256)
         else:
             return None
-
-
-class CryptoAction(enum.Enum):
-    Encrypt = 1
-    Decrypt = 2
-
-
-class CryptoOffload(blobxfer.models.offload._MultiprocessOffload):
-    def __init__(self, num_workers):
-        # type: (CryptoOffload, int) -> None
-        """Ctor for Crypto Offload
-        :param CryptoOffload self: this
-        :param int num_workers: number of worker processes
-        """
-        super(CryptoOffload, self).__init__(
-            self._worker_process, num_workers, 'Crypto')
-
-    def _worker_process(self):
-        # type: (CryptoOffload) -> None
-        """Crypto worker
-        :param CryptoOffload self: this
-        """
-        while not self.terminated:
-            try:
-                inst = self._task_queue.get(True, 1)
-            except queue.Empty:
-                continue
-            if inst[0] == CryptoAction.Encrypt:
-                # TODO on upload
-                raise NotImplementedError()
-            elif inst[0] == CryptoAction.Decrypt:
-                final_path, offsets, symkey, iv, encdata = \
-                    inst[1], inst[2], inst[3], inst[4], inst[5]
-                data = blobxfer.operations.crypto.aes_cbc_decrypt_data(
-                    symkey, iv, encdata, offsets.unpad)
-            self._done_cv.acquire()
-            self._done_queue.put((final_path, offsets, data))
-            self._done_cv.notify()
-            self._done_cv.release()
-
-    def add_decrypt_chunk(
-            self, final_path, offsets, symkey, iv, encdata):
-        # type: (CryptoOffload, str, blobxfer.models.download.Offsets,
-        #        bytes, bytes, bytes) -> None
-        """Add a chunk to decrypt
-        :param CryptoOffload self: this
-        :param str final_path: final path
-        :param blobxfer.models.download.Offsets offsets: offsets
-        :param bytes symkey: symmetric key
-        :param bytes iv: initialization vector
-        :param bytes encdata: encrypted data
-        """
-        self._task_queue.put(
-            (CryptoAction.Decrypt, final_path, offsets, symkey, iv,
-             encdata)
-        )
