+++ blobxfer/operations/azure/blob/page.py
@@ -40,9 +40,9 @@ logger = logging.getLogger(__name__)
 
 
 def create_client(storage_account):
-    # type: (blobxfer.models.azure.StorageAccount) -> PageBlobService
+    # type: (blobxfer.operations.azure.StorageAccount) -> PageBlobService
     """Create block blob client
-    :param blobxfer.models.azure.StorageAccount storage_account:
+    :param blobxfer.operations.azure.StorageAccount storage_account:
         storage account
     :rtype: PageBlobService
     :return: block blob service client
