+++ blobxfer/operations/md5.py
@@ -30,8 +30,14 @@ from builtins import (  # noqa
 )
 # stdlib imports
 import logging
+try:
+    import queue
+except ImportError:  # noqa
+    import Queue as queue
 # non-stdlib imports
 # local imports
+import blobxfer.models.azure
+import blobxfer.models.offload
 import blobxfer.util
 
 # create logger
@@ -72,3 +78,49 @@ def compute_md5_for_data_asbase64(data):
     hasher = blobxfer.util.new_md5_hasher()
     hasher.update(data)
     return blobxfer.util.base64_encode_as_string(hasher.digest())
+
+
+class LocalFileMd5Offload(blobxfer.models.offload._MultiprocessOffload):
+    """LocalFileMd5Offload"""
+    def __init__(self, num_workers):
+        # type: (LocalFileMd5Offload, int) -> None
+        """Ctor for Local File Md5 Offload
+        :param LocalFileMd5Offload self: this
+        :param int num_workers: number of worker processes
+        """
+        super(LocalFileMd5Offload, self).__init__(
+            self._worker_process, num_workers, 'MD5')
+
+    def _worker_process(self):
+        # type: (LocalFileMd5Offload) -> None
+        """Compute MD5 for local file
+        :param LocalFileMd5Offload self: this
+        """
+        while not self.terminated:
+            try:
+                filename, remote_md5, pagealign = self._task_queue.get(True, 1)
+            except queue.Empty:
+                continue
+            md5 = blobxfer.operations.md5.compute_md5_for_file_asbase64(
+                filename, pagealign)
+            logger.debug('MD5: {} <L..R> {} {}'.format(
+                md5, remote_md5, filename))
+            self._done_cv.acquire()
+            self._done_queue.put((filename, md5 == remote_md5))
+            self._done_cv.notify()
+            self._done_cv.release()
+
+    def add_localfile_for_md5_check(self, filename, remote_md5, mode):
+        # type: (LocalFileMd5Offload, str, str,
+        #        blobxfer.models.azure.StorageModes) -> None
+        """Add a local file to MD5 check queue
+        :param LocalFileMd5Offload self: this
+        :param str filename: file to compute MD5 for
+        :param str remote_md5: remote MD5 to compare against
+        :param blobxfer.models.azure.StorageModes mode: mode
+        """
+        if mode == blobxfer.models.azure.StorageModes.Page:
+            pagealign = True
+        else:
+            pagealign = False
+        self._task_queue.put((filename, remote_md5, pagealign))
