+++ test_requirements.txt
@@ -1,5 +1,5 @@
-flake8>=3.2.1
+flake8>=3.3.0
 mock>=2.0.0
 pypandoc>=1.3.3
-pytest>=3.0.5
+pytest>=3.0.7
 pytest-cov>=2.4.0
