+++ tests/test_blobxfer_models_download.py
@@ -13,10 +13,77 @@ except ImportError:  # noqa
 # non-stdlib imports
 import pytest
 # local imports
-import blobxfer.models
+import blobxfer.models.azure as azmodels
+import blobxfer.models.options as options
+import blobxfer.operations.azure as azops
 import blobxfer.util as util
 # module under test
-import blobxfer.download.models as models
+import blobxfer.models.download as models
+
+
+def test_localdestinationpath(tmpdir):
+    tmpdir.mkdir('1')
+    path = tmpdir.join('1')
+
+    a = models.LocalDestinationPath(str(path))
+    a.is_dir = True
+    assert str(a.path) == str(path)
+    assert a.is_dir
+
+    a.ensure_path_exists()
+    assert os.path.exists(str(a.path))
+
+    b = models.LocalDestinationPath()
+    b.is_dir = False
+    b.path = str(path)
+    with pytest.raises(RuntimeError):
+        b.ensure_path_exists()
+    assert not b.is_dir
+
+    path2 = tmpdir.join('2')
+    path3 = path2.join('3')
+    c = models.LocalDestinationPath(str(path3))
+    with pytest.raises(RuntimeError):
+        c.ensure_path_exists()
+    c.is_dir = False
+    c.ensure_path_exists()
+    assert os.path.exists(str(path2))
+    assert os.path.isdir(str(path2))
+    assert not c.is_dir
+
+
+def test_downloadspecification():
+    ds = models.Specification(
+        download_options=options.Download(
+            check_file_md5=True,
+            chunk_size_bytes=4194304,
+            delete_extraneous_destination=False,
+            mode=azmodels.StorageModes.Auto,
+            overwrite=True,
+            recursive=True,
+            restore_file_attributes=False,
+            rsa_private_key=None,
+        ),
+        skip_on_options=options.SkipOn(
+            filesize_match=True,
+            lmt_ge=False,
+            md5_match=True,
+        ),
+        local_destination_path=models.LocalDestinationPath('dest'),
+    )
+
+    asp = azops.SourcePath()
+    p = 'some/remote/path'
+    asp.add_path_with_storage_account(p, 'sa')
+
+    ds.add_azure_source_path(asp)
+
+    assert ds.options.check_file_md5
+    assert not ds.skip_on.lmt_ge
+    assert ds.destination.path == pathlib.Path('dest')
+    assert len(ds.sources) == 1
+    assert p in ds.sources[0]._path_map
+    assert ds.sources[0]._path_map[p] == 'sa'
 
 
 def test_downloaddescriptor(tmpdir):
@@ -25,14 +92,14 @@ def test_downloaddescriptor(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = True
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 1024
     ase._encryption = mock.MagicMock()
     with pytest.raises(RuntimeError):
-        d = models.DownloadDescriptor(lp, ase, opts)
+        d = models.Descriptor(lp, ase, opts)
 
     ase._encryption.symmetric_key = b'123'
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     assert d.entity == ase
     assert not d.must_compute_md5
@@ -44,19 +111,19 @@ def test_downloaddescriptor(tmpdir):
 
     d.local_path.unlink()
     ase._size = 1
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     assert d._total_chunks == 1
     assert d.local_path.stat().st_size == 0
 
     d.local_path.unlink()
     ase._encryption = None
     ase._size = 1024
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     assert d.local_path.stat().st_size == 1024
 
     # pre-existing file check
     ase._size = 0
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     assert d._total_chunks == 0
     assert d.local_path.stat().st_size == 0
 
@@ -67,9 +134,9 @@ def test_downloaddescriptor_next_offsets(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = True
     opts.chunk_size_bytes = 256
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 128
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     offsets = d.next_offsets()
     assert d._total_chunks == 1
@@ -82,12 +149,12 @@ def test_downloaddescriptor_next_offsets(tmpdir):
     assert d.next_offsets() is None
 
     ase._size = 0
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     assert d._total_chunks == 0
     assert d.next_offsets() is None
 
     ase._size = 1
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     offsets = d.next_offsets()
     assert d._total_chunks == 1
     assert offsets.chunk_num == 0
@@ -99,7 +166,7 @@ def test_downloaddescriptor_next_offsets(tmpdir):
     assert d.next_offsets() is None
 
     ase._size = 256
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     offsets = d.next_offsets()
     assert d._total_chunks == 1
     assert offsets.chunk_num == 0
@@ -111,7 +178,7 @@ def test_downloaddescriptor_next_offsets(tmpdir):
     assert d.next_offsets() is None
 
     ase._size = 256 + 16
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     offsets = d.next_offsets()
     assert d._total_chunks == 2
     assert offsets.chunk_num == 0
@@ -132,7 +199,7 @@ def test_downloaddescriptor_next_offsets(tmpdir):
     ase._encryption = mock.MagicMock()
     ase._encryption.symmetric_key = b'123'
     ase._size = 128
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     offsets = d.next_offsets()
     assert d._total_chunks == 1
     assert offsets.chunk_num == 0
@@ -144,7 +211,7 @@ def test_downloaddescriptor_next_offsets(tmpdir):
     assert d.next_offsets() is None
 
     ase._size = 256
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     offsets = d.next_offsets()
     assert d._total_chunks == 1
     assert offsets.chunk_num == 0
@@ -156,7 +223,7 @@ def test_downloaddescriptor_next_offsets(tmpdir):
     assert d.next_offsets() is None
 
     ase._size = 256 + 32  # 16 bytes over + padding
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     offsets = d.next_offsets()
     assert d._total_chunks == 2
     assert offsets.chunk_num == 0
@@ -181,9 +248,9 @@ def test_postpone_integrity_check(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = True
     opts.chunk_size_bytes = 32
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     offsets = d.next_offsets()
     d._postpone_integrity_check(offsets, b'0' * ase._size)
@@ -198,9 +265,9 @@ def test_postpone_integrity_check(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = False
     opts.chunk_size_bytes = 32
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     offsets = d.next_offsets()
     d._postpone_integrity_check(offsets, b'0' * ase._size)
@@ -219,9 +286,9 @@ def test_perform_chunked_integrity_check(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = True
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     offsets = d.next_offsets()
     data = b'0' * opts.chunk_size_bytes
@@ -233,11 +300,11 @@ def test_perform_chunked_integrity_check(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = False
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
     ase._encryption = mock.MagicMock()
     ase._encryption.symmetric_key = b'123'
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     offsets = d.next_offsets()
     data = b'0' * opts.chunk_size_bytes
@@ -255,10 +322,10 @@ def test_cleanup_all_temporary_files(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = False
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 16
     lp = pathlib.Path(str(tmpdir.join('a')))
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     offsets = d.next_offsets()
     data = b'0' * opts.chunk_size_bytes
@@ -269,7 +336,7 @@ def test_cleanup_all_temporary_files(tmpdir):
     assert not d._unchecked_chunks[0].file_path.exists()
 
     lp = pathlib.Path(str(tmpdir.join('b')))
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     offsets = d.next_offsets()
     data = b'0' * opts.chunk_size_bytes
@@ -288,9 +355,9 @@ def test_write_data(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = True
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
 
     offsets = d.next_offsets()
     data = b'0' * ase._size
@@ -306,7 +373,7 @@ def test_finalize_file(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = False
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
     ase._encryption = mock.MagicMock()
     ase._encryption.symmetric_key = b'123'
@@ -322,7 +389,7 @@ def test_finalize_file(tmpdir):
         message_authentication_code = util.base64_encode_as_string(
             _hmac.digest())
 
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     d.hmac.update(data)
     d.finalize_file()
 
@@ -335,7 +402,7 @@ def test_finalize_file(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = True
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
 
     data = b'0' * ase._size
@@ -343,7 +410,7 @@ def test_finalize_file(tmpdir):
     md5.update(data)
     ase._md5 = util.base64_encode_as_string(md5.digest())
 
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     d.md5.update(data)
     d.finalize_file()
 
@@ -356,12 +423,12 @@ def test_finalize_file(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = False
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
 
     data = b'0' * ase._size
 
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     d.finalize_file()
 
     assert not d.local_path.exists()
@@ -373,13 +440,13 @@ def test_finalize_file(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = True
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
 
     data = b'0' * ase._size
     ase._md5 = 'oops'
 
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     d.md5.update(data)
     d.finalize_file()
 
@@ -392,10 +459,10 @@ def test_operations(tmpdir):
     opts = mock.MagicMock()
     opts.check_file_md5 = True
     opts.chunk_size_bytes = 16
-    ase = blobxfer.models.AzureStorageEntity('cont')
+    ase = azmodels.StorageEntity('cont')
     ase._size = 32
 
-    d = models.DownloadDescriptor(lp, ase, opts)
+    d = models.Descriptor(lp, ase, opts)
     d._outstanding_ops = 1
     d._unchecked_chunks = {0: None}
     assert not d.all_operations_completed
