+++ tests/test_blobxfer_operations_azure_blob_append.py
@@ -1,29 +1,28 @@
 # coding=utf-8
-"""Tests for models"""
+"""Tests for operations: blob append"""
 
 # stdlib imports
 # non-stdlib imports
 import azure.storage
-import pytest
 # local imports
-import blobxfer.models as models
 # module under test
-import blobxfer.blob.block.operations as ops
+import blobxfer.operations.azure as azops
+import blobxfer.operations.azure.blob.append as ops
 
 
 def test_create_client():
-    sa = models.AzureStorageAccount('name', 'key', 'endpoint')
+    sa = azops.StorageAccount('name', 'key', 'endpoint')
     client = ops.create_client(sa)
     assert client is not None
-    assert isinstance(client, azure.storage.blob.BlockBlobService)
+    assert isinstance(client, azure.storage.blob.AppendBlobService)
     assert isinstance(
         client.authentication,
         azure.storage._auth._StorageSharedKeyAuthentication)
 
-    sa = models.AzureStorageAccount('name', '?key&sig=key', 'endpoint')
+    sa = azops.StorageAccount('name', '?key&sig=key', 'endpoint')
     client = ops.create_client(sa)
     assert client is not None
-    assert isinstance(client, azure.storage.blob.BlockBlobService)
+    assert isinstance(client, azure.storage.blob.AppendBlobService)
     assert isinstance(
         client.authentication,
         azure.storage._auth._StorageSASAuthentication)
