+++ tests/test_blobxfer_operations_azure_blob_page.py
@@ -5,24 +5,24 @@
 # non-stdlib imports
 import azure.storage
 # local imports
-import blobxfer.models as models
 # module under test
-import blobxfer.blob.append.operations as ops
+import blobxfer.operations.azure as azops
+import blobxfer.operations.azure.blob.page as ops
 
 
 def test_create_client():
-    sa = models.AzureStorageAccount('name', 'key', 'endpoint')
+    sa = azops.StorageAccount('name', 'key', 'endpoint')
     client = ops.create_client(sa)
     assert client is not None
-    assert isinstance(client, azure.storage.blob.AppendBlobService)
+    assert isinstance(client, azure.storage.blob.PageBlobService)
     assert isinstance(
         client.authentication,
         azure.storage._auth._StorageSharedKeyAuthentication)
 
-    sa = models.AzureStorageAccount('name', '?key&sig=key', 'endpoint')
+    sa = azops.StorageAccount('name', '?key&sig=key', 'endpoint')
     client = ops.create_client(sa)
     assert client is not None
-    assert isinstance(client, azure.storage.blob.AppendBlobService)
+    assert isinstance(client, azure.storage.blob.PageBlobService)
     assert isinstance(
         client.authentication,
         azure.storage._auth._StorageSASAuthentication)
